progress,tq: move *progress.Meter to *tq.Meter
