+++ commands/command_checkout.go
@@ -7,8 +7,8 @@ import (
 	"github.com/git-lfs/git-lfs/filepathfilter"
 	"github.com/git-lfs/git-lfs/git"
 	"github.com/git-lfs/git-lfs/lfs"
-	"github.com/git-lfs/git-lfs/progress"
 	"github.com/git-lfs/git-lfs/tasklog"
+	"github.com/git-lfs/git-lfs/tq"
 	"github.com/spf13/cobra"
 )
 
@@ -28,7 +28,7 @@ func checkoutCommand(cmd *cobra.Command, args []string) {
 	var totalBytes int64
 	var pointers []*lfs.WrappedPointer
 	logger := tasklog.NewLogger(os.Stdout)
-	meter := progress.NewMeter(&progress.MeterOption{
+	meter := tq.NewMeter(&tq.MeterOption{
 		OS: cfg.Os,
 	})
 	logger.Enqueue(meter)
