+++ commands/uploader.go
@@ -12,7 +12,6 @@ import (
 
 	"github.com/git-lfs/git-lfs/errors"
 	"github.com/git-lfs/git-lfs/lfs"
-	"github.com/git-lfs/git-lfs/progress"
 	"github.com/git-lfs/git-lfs/tasklog"
 	"github.com/git-lfs/git-lfs/tools"
 	"github.com/git-lfs/git-lfs/tq"
@@ -58,7 +57,7 @@ type uploadContext struct {
 	gitfilter    *lfs.GitFilter
 
 	logger *tasklog.Logger
-	meter  *progress.Meter
+	meter  *tq.Meter
 	tq     *tq.TransferQueue
 
 	committerName  string
