+++ tq/transfer_queue.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/git-lfs/git-lfs/errors"
 	"github.com/git-lfs/git-lfs/lfsapi"
-	"github.com/git-lfs/git-lfs/progress"
 	"github.com/git-lfs/git-lfs/tools"
 	"github.com/rubyist/tracerx"
 )
@@ -109,7 +108,7 @@ type TransferQueue struct {
 	adapterInitMutex  sync.Mutex
 	dryRun            bool
 	cb                tools.CopyCallback
-	meter             *progress.Meter
+	meter             *Meter
 	errors            []error
 	transfers         map[string]*objects
 	batchSize         int
@@ -178,7 +177,7 @@ func DryRun(dryRun bool) Option {
 	}
 }
 
-func WithProgress(m *progress.Meter) Option {
+func WithProgress(m *Meter) Option {
 	return func(tq *TransferQueue) {
 		tq.meter = m
 	}
